Fix bug that shows expired non-boosted proposals as passed
If they have more yes votes than no
