+++ src/reducers/arcReducer.ts
@@ -1,8 +1,10 @@
-import * as ActionTypes from "constants/arcConstants";
+import { ExecutionState } from "@daostack/arc.js";
 import * as update from "immutability-helper";
+import * as moment from "moment";
+
+import * as ActionTypes from "constants/arcConstants";
 import { RedeemAction, StakeAction, VoteAction, CreateProposalAction } from "actions/arcActions";
 import { AsyncActionSequence } from "actions/async";
-import { ExecutionState } from "@daostack/arc.js";
 
 export enum ProposalStates {
   None = 0,
@@ -148,6 +150,39 @@ export const initialState: IArcState = {
   daos: {},
   proposals: {},
 };
+
+export function proposalEnded(proposal: IProposalState) {
+  const res = (
+    proposal.state == ProposalStates.Executed ||
+    proposal.state == ProposalStates.Closed ||
+    // Boosted proposal past end time but not yet executed
+    (proposal.state == ProposalStates.Boosted && proposal.boostedTime + proposal.boostedVotePeriodLimit <= +moment() / 1000) ||
+    // Pre boosted proposal past end time but not yet executed
+    (proposal.state == ProposalStates.PreBoosted && proposal.submittedTime + proposal.preBoostedVotePeriodLimit <= +moment() / 1000)
+  );
+  return res;
+}
+
+export function proposalPassed(proposal: IProposalState) {
+  const res = (
+    (proposal.state == ProposalStates.Executed && proposal.winningVote == VoteOptions.Yes) ||
+    // Boosted proposal past end time with more yes votes than no, but not yet executed
+    (proposal.state == ProposalStates.Boosted && proposal.boostedTime + proposal.boostedVotePeriodLimit <= +moment() / 1000 && proposal.winningVote == VoteOptions.Yes)
+  );
+  return res;
+}
+
+export function proposalFailed(proposal: IProposalState) {
+  const res = (
+    proposal.state == ProposalStates.Closed ||
+    (proposal.state == ProposalStates.Executed && proposal.winningVote == VoteOptions.No) ||
+    // Boosted proposal past end time with more no votes than yes, but not yet executed
+    (proposal.state == ProposalStates.Boosted && proposal.boostedTime + proposal.boostedVotePeriodLimit <= +moment() / 1000 && proposal.winningVote == VoteOptions.No) ||
+    // Pre boosted proposal past end time but not yet executed are always failed
+    (proposal.state == ProposalStates.PreBoosted && proposal.submittedTime + proposal.preBoostedVotePeriodLimit <= +moment() / 1000)
+  );
+  return res;
+}
 
 const arcReducer = (state = initialState, action: any) => {
   const { payload } = action;
